A66
This code uses reference variable of min and max in order to get the values of both min and max from the single function. The function itself uses 4 if statements, 2 for min and 2 for max. By default, the min and max is set a and then the if statements compares if the new max is the largest number and if the new min is the smallest number. It does the same comparison for b and c. Since the values min and max are in reference variables, there isn't a need to end the program with a return statement.